1. Post id int -> Long타입으로 변경 2. userId 참조값이 userProvideId (String타입)이므로 맞게 변경

+++ src/main/java/com/example/demo/src/hanamoa/controller/PostController.java
@@ -6,8 +6,6 @@ import com.example.demo.src.hanamoa.service.PostService;
 import com.example.demo.common.response.BaseResponse;
 import com.example.demo.common.response.BaseResponseStatus;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -20,6 +18,7 @@ public class PostController {
     @Autowired
     private PostService postService;
 
+
     // 모든 게시글을 가져오는 API
     @GetMapping
     public BaseResponse<List<PostResponse>> getAllPosts() {
@@ -29,7 +28,7 @@ public class PostController {
 
     // 특정 게시글을 ID로 조회하는 API
     @GetMapping("/{id}")
-    public BaseResponse<PostResponse> getPostById(@PathVariable int id) {
+    public BaseResponse<PostResponse> getPostById(@PathVariable Long id) {
         PostResponse post = postService.getPostById(id); // ID로 게시글 조회
         return new BaseResponse<>(post); // 성공 응답 반환
     }
@@ -44,14 +43,14 @@ public class PostController {
     // 특정 게시글을 수정하는 API
     @Operation(summary = "게시글 수정",description = "게시글 상세 정보를 JSON으로 받아 수정")
     @PutMapping("/{id}")
-    public BaseResponse<String> updatePost(@PathVariable int id, @RequestBody PostRequest postRequest) {
+    public BaseResponse<String> updatePost(@PathVariable Long id, @RequestBody PostRequest postRequest) {
         postService.updatePost(id, postRequest); // 게시글 수정
         return new BaseResponse<>(BaseResponseStatus.SUCCESS); // 성공 응답 반환
     }
 
     // 특정 게시글을 삭제하는 API
     @DeleteMapping("/{id}")
-    public BaseResponse<String> deletePost(@PathVariable int id) {
+    public BaseResponse<String> deletePost(@PathVariable Long id) {
         postService.deletePost(id); // 게시글 삭제
         return new BaseResponse<>(BaseResponseStatus.SUCCESS); // 성공 응답 반환
     }
